pyproject.toml
[project]
name = "backloganalyzer"
version = "0.0.3"
description = "Answer questions about your GitHub backlog"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
   "joblib",
   "openai",
   "PyGitHub",
   "numpy",
   "tqdm"
]
[project.scripts]
banal = "banal:main"
